// src/FE/src/api/generated/models/documentDetailResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Workslip.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface DocumentDetailResponse {
  id: string;
  title: string;
  content: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
  /** @nullable */
  createdByUserId: string | null;
  /** @nullable */
  createdByDisplayName: string | null;
  /** @nullable */
  updatedByUserId: string | null;
  /** @nullable */
  updatedByDisplayName: string | null;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  revision: number | string;
}